SQL query for gender and Gravity Forms
Using an SQL query to pull a single variable from the wordpress database and populate a field in gravity forms.  The field must be named 'gender'.

This pulls data from the data table created after a user submits their statistics for the first time as the gender is recorded in that table, it then links the gender to the new form so that the user does not have to input the gender each time a form is submitted.

This is inserted using the 'code snipppets' plugin by sheabunge (respository at https://github.com/sheabunge/code-snippets) but if not using that plugin, must be inserted into the functions.php file in wordpress.